Going on a warning hunt!
Got rid of a lot of compiler warnings, and found several bugs related to
type conversion in the process.

+++ include/AppStateGameMode.h
@@ -36,7 +36,7 @@ class AppStateGameMode : public State
     EventResult handle_key_press(sf::Event::KeyEvent& key);
     EventResult handle_mouse_wheel(sf::Event::MouseWheelEvent& wheel);
 
-    void add_zoom(double zoom_amount);
+    void add_zoom(float zoom_amount);
 
   private:
     struct Impl;
